requirements.txt
uvicorn[standard]
motor
pydantic-settings
python-dotenv
fastapi
google-genai
Pillow
httpx
pandas